Optimize statistics computation by using getblockstats
Previously the code used `getblock` with verbosity 2 (includes all TX
details), which is a very expensive call for larger blocks. Because we
only get the number of inputs/outputs/total value from this call, this
is wasteful as the same can be computed on the bitcoind side using the
`getblockstats` RPC call, which transfers only a minimal amount of data
and uses much less memory.

Additionally we cache the result of `getblockstats` until we need stats
from a different block.